gcpdeploy/streamlit/.ipynb_checkpoints/requirements-checkpoint.txt
streamlit
google-cloud-aiplatform
